requirements.txt
disnake[speed,voice] == 2.9.0
python-dotenv == 1.0.0